still working on custom fonts

+++ Eve2Display.h
@@ -20,7 +20,7 @@ class Eve2Display {
     /* debug */
     void test();
     void printRAM_DL();
-    void printRAM(uint32_t address, uint16_t length);
+    void printRAM(uint32_t address, uint32_t length);
     void printCommands();
     void log(uint8_t level, char *msg, uint32_t v);
